Store step-1 price and bar OHLC as int32
Prices are whole ticks (x4), and any ES price fits int32 (int16 would wrap
above 32,767 ticks = 8,191.75 points). Step 1 writes price as int32; to_dat
widens it to int64 for tick.dat (byte-identical) and writes bar
open/high/low/close as int32. Volume stays int64. Values are unchanged;
parquet sizes barely move (zstd already packed the int64s), memory halves
for these columns when loaded. md updated.

+++ raw_data_preprocessing.py
@@ -143,8 +143,9 @@ def to_ticks(table, session_date, pdt_code):
 
     df = pd.DataFrame({'ts': ts, 'price': df['price'], 'size': df['size'].astype('int64')})
     # ES trades in 0.25-point ticks; price x4 is the price in ticks, an exact integer (4000.25 -> 16001).
-    # OHLC downstream is built from it, so every price column is in ticks.
-    df['price'] = (df['price'] * 4).astype('int64')
+    # OHLC downstream is built from it, so every price column is in ticks. int32 holds any ES price
+    # (int16 would wrap above 32,767 ticks = 8,191.75 points); to_dat.py widens it to int64 for tick.dat.
+    df['price'] = (df['price'] * 4).astype('int32')
 
     out = df.groupby(['ts', 'price'], sort=False).agg(volume=('size', 'sum')).reset_index()
     out['pdt_code'] = pdt_code
